requirements.txt
Django==3.2.4
djangorestframework==3.12.4
django-filter==2.4.0
python-dotenv==0.17.1